# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the online Book Reservation service.",
        "title": "Book Reservation API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "paths": {
        "/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new User",
                "parameters": [
                    {
                        "description": "User details",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RegisterRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.RegisterResponseDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HttpError"
                        }
                    }
                }
            }
        },
        "/users": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create an User",
                "parameters": [
                    {
                        "description": "User details",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.CreateUserRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.CreateUserResponseDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HttpError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.JwtToken": {
            "type": "object",
            "properties": {
                "raw": {
                    "type": "string"
                },
                "signed": {
                    "type": "string"
                }
            }
        },
        "auth.RegisterRequestDTO": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string",
                    "example": "2024-08-15"
                },
                "cpf": {
                    "type": "string",
                    "example": "32212276723"
                },
                "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                }
            }
        },
        "auth.RegisterResponseDTO": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "$ref": "#/definitions/auth.JwtToken"
                },
                "user": {
                    "$ref": "#/definitions/auth.RegisterUser"
                }
            }
        },
        "auth.RegisterUser": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "cpf": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "user.CreateUserRequestDTO": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "cpf": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "description": "@example {\"name\": \"John Doe\", \"email\": \"john.doe@example.com\", \"password\": \"teste1234\", \"cpf\": \"23212332112\", \"birth_date\": \"2024-08-12\"}",
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "user.CreateUserResponseDTO": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "cpf": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "utils.HttpError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}